+++ src/main/kotlin/domain/InputManager.kt
@@ -4,5 +4,5 @@ interface InputManager {
     fun showPrompt(prompt: String)
     fun getString() : String
     fun getInt() : Int
-    fun getFloat() : Float
+    fun getDouble() : Double
 }
